requirements.txt
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Django==4.2.4
fastapi==0.100.1
filelock==3.12.2
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.42.0
gevent==23.7.0
geventhttpclient==2.0.9
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
locust==2.15.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rich==13.5.2
roundrobin==0.0.4
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
streamlit==1.25.0
streamlit-aggrid==0.3.4.post3
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.6
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
